Revert "fix: another attempt on long pids"
This reverts commit 6ba0d92dd600b2fdac0200c3c24959897ab33f8e.

+++ common/pid.h
@@ -22,8 +22,6 @@
               2014 Vitaly Valtman
 */
 
-#include <stdint.h>
-
 #pragma once
 
 #pragma pack(push,4)
@@ -31,14 +29,14 @@
 struct process_id {
   unsigned ip;
   short port;
-  uint32_t pid;
+  unsigned short pid;
   int utime;
 };
 
 struct process_id_ext {
   unsigned ip;
   short port;
-  uint32_t pid;
+  unsigned short pid;
   int utime;
   int actor_id;
 };
